bindings: Do not install python to absolute address
When CMake's install encounters an absolute path, it will ignore the
INSTALL_PREFIX the user can provide. Add the prefix to the install-path
so that we can override the destination if we so choose.

This is useful when running the installer inside a docker-container as
we do not run this as root.