Split emulators out from libi3ds and compile as separate library
A lot of the code currently included in libi3ds belongs to the
emulators. These are laregly used by the test infrastructure and should
not clutter the library we use for the ordinary service.

By splitting these out in a separate library we create a better split
between test and production and we also reduce the size of libi3ds.so
from 13MB down to 8.8.

Further work to reduce the size of i3ds is ongoing.