Clean up CXX_COMPILER_FLAGS and expose DEBUG as option
In commit 15b91e55aab ("Add valgrind to build system"), -ggdb3 was added
to give valgrind symbols to work with. An unforeseen consequence of
this, was the massive increase in size of libi3ds.so, a near 10-fold
growth in size.

In the same series of commits, the old DEBUG-switch was broken as
options were added in the wrong place. This moves the CXX_FLAGS to the
correct place and also exposes BUILD_DEBUG as an option (default OFF).

In debug builds:
libi3ds.so: 12 MM
libi3ds_emulators: 6.3 MB

In normal build:
libi3ds.so: 857 kB
libi3ds_emulators: 402 kB