Adjust NodeID for TIR, HR and Stereo camera in test_recording
In commit 5751dd819a54 "Adds emulated trigger and corresponding tests" 2
new nodes were added and exposed through the emulator_factory API.

Then, in commit 898341245601 "Fixes for triggers of suite emulator",
both trigger and flash was moved to an earlier stage in the setup, and
this gave the emulated TIR, HR and stereo-camera new NodeIDs.

This would then cause test_recording to attempt to activate (and
deactivate) the emulated trigger and flash, an operation neither
supported. Since test_recording did not check for failure here, this
failed silently.

Change the array of NodeIDs for TIR, HR and stereo to properly activate
the cameras. This will probably fail later as this approach is a bit
fragile. The test, due to the previous update to test_recording.sh, will
fail however, so we should catch it a bit sooner the next time.