test: make _i3ds_binding.so available to test and set PYTHONPATH correctly
Referencing _i3ds_binding.so is only possible after 'make install', and
making a copy of _i3ds_python.so requires a build-target as CMake cannot
find this file when generating the enviornment.

So, create a custom command that creates the copy *after* the .so has
been built/compiled and fix the typo in PYTHONPATH in test/

Finally, to track down issues later, dump the defined path from the
python script if import fails.