CMakeList: disalbe unused-parameter warning
Files generated by asn1cc will use a template to create the parsers. For
some of the functions, not all parameters are useful, but this is a bit
complicated to avoid in C with a semi-simple parser.

Instead of manually fixing these files upon very new revision of
generated/, disable this warning. Since this is included directly into
libi3ds, this means that we pretty much have to disable this globally.

Cc: Fredrik Bakkevig Haugli <fredrik.haugli@sintef.no>