python: provide a flag to show that we actually build bindings
It is not enough to guard the python-bindings test with BUILD_BINDINGS
only as bindings will abort if it is missing other dependencies. We
therefore add a flag in bindings/ to indicate that we are building it.

This should avoid situations where a warning in cmake is missed and the
python-bindings test fail due to missing i3ds_bindings.

Reported-by: Fredrik Bakkevig Haugli <fredrik.haugli@sintef.no>